IHM/Chat Streamlit/requirements.txt
streamlit
google-generativeai
python-dotenv
speechrecognition
pillow